Isolate Sinopia server operations to a single module
This is a refactoring commit with the following pieces included:
* Remove knowledge of Sinopia server internals, including authentication, from `ImportResourceTemplate` component
* Extract `sinopiaServerSpoof` module from `sinopiaServer` to disentangle them
* Drop unnecessary `.js[x]` extension from `import` statements for consistency
