+++ src/components/NewsPanel.jsx
@@ -1,7 +1,7 @@
 // Copyright 2018 Stanford Junior University see Apache2.txt for license
 
 import React, {Component} from 'react'
-import NewsItem from './NewsItem.jsx'
+import NewsItem from './NewsItem'
 
 class NewsPanel extends Component {
   constructor(props) {
